setup.cfg: list explicit license file
This is not needed for wheels, since the default license files pattern
already finds the current name of this file. However, it is needed for
sdist in order to guarantee the file gets disted.

Currently, that file gets disted anyway... but only because MANIFEST.in
forces it to be.

So adding the license_files metadata lets us drop the info from
MANIFEST.in; however, as the project maintainer finds it easier to look
in only one file for this information, this is being left in despite its
redundancy.